requirements.txt
torch>=2.0
torchvision>=0.15
h5py>=3.8
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
matplotlib>=3.7
scikit-image>=0.21
tqdm>=4.65
tensorboard>=2.13
einops>=0.6
